usbarmory-commandsubclient/python/requirements.txt
google-auth-oauthlib
google-cloud-pubsub
jwt